+++ rbxmk/testdata/fsaccess.lua
@@ -1,14 +1,14 @@
 T.Fail(function() return fs.stat("/") ~= nil end, "fs root")
 
-T.Fail(function() return fs.stat(os.expand("$wd")) ~= nil end, "stat working directory")
+T.Fail(function() return fs.stat(path.expand("$wd")) ~= nil end, "stat working directory")
 T.Fail(function() return fs.stat(".") ~= nil end, "stat working directory w/ dot")
 T.Pass(function() return fs.dir(".") ~= nil end, "dir working directory")
 
-T.Pass(function() return fs.stat(os.expand("$sd")) ~= nil end, "script directory")
+T.Pass(function() return fs.stat(path.expand("$sd")) ~= nil end, "script directory")
 
-T.Pass(function() return fs.stat(os.expand("$rsd")) ~= nil end, "root script directory")
+T.Pass(function() return fs.stat(path.expand("$rsd")) ~= nil end, "root script directory")
 
-T.Pass(function() return fs.stat(os.expand("$wd/foo/bar/baz")) == nil end, "nonextant")
+T.Pass(function() return fs.stat(path.expand("$wd/foo/bar/baz")) == nil end, "nonextant")
 
 T.Fail(function() return fs.stat("..") == nil end, "wd parent")
 T.Fail(function() return fs.stat("../..") == nil end, "wd ancestor")