+++ rbxmk/testdata/roblox/Instance/desc_inherit.lua
@@ -1,5 +1,5 @@
-local a = fs.read(os.expand("$sd/../../dump.desc.json"))
-local b = fs.read(os.expand("$sd/../../dump.desc.json"))
+local a = fs.read(path.expand("$sd/../../dump.desc.json"))
+local b = fs.read(path.expand("$sd/../../dump.desc.json"))
 local n = nil
 local f = false